+++ src/app/game/list/page.tsx
@@ -7,9 +7,10 @@ import CustomPlayerType from './components/CustomPlayerType';
 import CustomGenres from './components/CustomGenres';
 import typeConverter from '@/utils/typeConverter';
 import CustomResetComponent from './components/CustomResetComponent';
-import { Suspense } from 'react';
-import { Skeleton } from '@/components/ui/skeleton';
 import SuspendedGameDisplay from './components/SuspendedGameDisplay';
+import { getQueryClient } from '@/hooks/getQueryClient';
+import { getGames } from './getGames';
+import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 
 const imageList = [
   { title: 'Bioshock: Infinite', img_src: '/img/hero/bg_gameList_1.webp' },
@@ -51,20 +52,17 @@ export default async function GameList({
     size: 12,
     sort: [],
   };
-  const PickCardSkeletons = () => {
-    return (
-      <div className="lg:w-[1280px] grid grid-cols-4 grid-rows-3 gap-x-6 gap-y-12 mt-[100px] mb-[100px]">
-        {Array.from({ length: 12 }).map((_, ind) => (
-          <Skeleton className="w-full aspect-square rounded-full" key={`Skeleton_Games_${ind}`} />
-        ))}
-      </div>
-    );
-  };
+  const queryClient = getQueryClient();
+  await queryClient.prefetchQuery({
+    queryKey: ['games', step, pagination],
+    queryFn: async () => getGames(step, pagination),
+  });
 
   function formatDateStr(dateStr: string) {
     const temp = dateStr.slice(0, dateStr.length - 14) + '+' + dateStr.slice(dateStr.length - 13);
     return temp;
   }
+
   return (
     <div className="flex flex-col items-center">
       <section className="w-full h-[520px]">
@@ -125,9 +123,9 @@ export default async function GameList({
           </div>
         </div>
       </div>
-      <Suspense fallback={<PickCardSkeletons />}>
+      <HydrationBoundary state={dehydrate(queryClient)}>
         <SuspendedGameDisplay pagination={pagination} step={step} />
-      </Suspense>
+      </HydrationBoundary>
     </div>
   );
 }
