+++ src/hooks/useReactQuery.tsx
@@ -1,38 +1,10 @@
 'use client';
 
-// import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { QueryClientProvider } from '@tanstack/react-query';
+import { getQueryClient } from './getQueryClient';
 
 export default function ReactQueryProviders({ children }: React.PropsWithChildren) {
-  function makeQueryClient() {
-    return new QueryClient({
-      defaultOptions: {
-        queries: {
-          staleTime: 60 * 1000,
-        },
-      },
-    });
-  }
-
-  let browserQueryClient: QueryClient | undefined = undefined;
-
-  function getQueryClient() {
-    if (typeof window === 'undefined') {
-      // Server일 경우
-      return makeQueryClient();
-    } else {
-      // Browser일 경우
-      if (!browserQueryClient) browserQueryClient = makeQueryClient();
-      return browserQueryClient;
-    }
-  }
-
   const queryClient = getQueryClient();
 
-  return (
-    <QueryClientProvider client={queryClient}>
-      {children}
-      {/* <ReactQueryDevtools initialIsOpen={false} /> */}
-    </QueryClientProvider>
-  );
+  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
 }
